LayoutTests/media/encrypted-media/mock-MediaKeys-createSession.html

<!DOCTYPE html>
<html>
<head>
    <script src=../video-test.js></script>
    <script type="text/javascript">
    var mock;
    var promise;
    var mediaKeySystemAccess;
    var mediaKeys;
    var mediaKeySession;
    var capabilities = {};

    function doTest()
    {
        if (!window.internals) {
            failTest("Internals is required for this test.")
            return;
        }

        run('internals.initializeMockMediaSource()');
        run('mock = internals.registerMockCDM()');
        run('mock.supportedDataTypes = ["mock"]');
        run('capabilities.initDataTypes = ["mock"]');
        run(`capabilities.videoCapabilities = [{ contentType: 'video/mock; codecs="mock"' }] `);
        run('promise = navigator.requestMediaKeySystemAccess("org.webkit.mock", [capabilities])');
        promise.then(gotMediaKeySystemAccess).catch(failTest);
    }

    function next() {
        if (!tests.length) {
            mock.unregister();
            endTest()
            return;
        }

        var nextTest = tests.shift();
        consoleWrite('');
        nextTest();
    }

    function shouldResolve(promise) {
        promise.then(mediaKeySystemAccess => {
            logResult(Success, 'Promise resolved');
            next();
        }, () => {
            logResult(Failed, 'Promise rejected');
            next();
        });
    }

    function shouldReject(promise) {
        promise.then(() => {
            logResult(Failed, 'Promise resolved incorrectly');
            next();
        }, exceptionCode => {
            logResult(Success, 'Promise rejected correctly');
            next();
        });
    }

    function gotMediaKeySystemAccess(result) {
        logResult(Success, 'Promise resolved');
        mediaKeySystemAccess = result;
        next();
    }

    function gotMediaKeys(result) {
        logResult(Success, 'Promise resolved');
        mediaKeys = result;
        next();
    }

    function stringToUInt8Array(str)
    {
       var array = new Uint8Array(str.length);
       for (var i=0; i<str.length; i++)
            array[i] = str.charCodeAt(i);
       return array;
    }

    tests = [
        function() {
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            promise.then(gotMediaKeys).catch(failTest);
        },

        function() {
            run('mediaKeySession = mediaKeys.createSession("temporary")');
            testExpected('typeof mediaKeySession', 'object')
            next();
        },

        function() {
            run('mock.supportedSessionTypes = ["temporary"]');
            testDOMException('mediaKeySession = mediaKeys.createSession("persistent-license")', 'DOMException.NOT_SUPPORTED_ERR');
            next();
        },

        function() {
            run('mock.supportsSessions = false');
            testDOMException('mediaKeySession = mediaKeys.createSession("temporary")', 'DOMException.INVALID_STATE_ERR');
            next();
        }
    ];
    </script>
</head>
<body onload="doTest()">
</body>
</html>
